+++ tests/testthat/helper-revdepcheck.R
@@ -1,0 +1,9 @@
+
+# From dev testthat
+skip_if_offline <- function(host = "r-project.org") {
+  skip_if_not_installed("curl")
+  has_internet <- !is.null(curl::nslookup(host, error = FALSE))
+  if (!has_internet) {
+    skip("offline")
+  }
+}